refactor: move serve functions code to utils file
